test(finalize): dedupe shard helpers, split tests onto entrypoint rails
Extract the verbatim-duplicated _write_minimal_wds_shard and the
finalize-specific wds/hdf5 DatasetSpec builders + shard seeders into a
focused tests/helpers/finalize_shards.py, shared by the entrypoint,
branch-unit, and real-R2 lanes so a shard-layout change updates every
caller at once.

Bring finalize onto the codified entrypoint rails the way generate_dataset
is split: the entrypoint-level finalize(cfg) / main() tests move to the
canonical tests/test_finalize_dataset.py, while the branch-level
finalize_wds / finalize_hdf5 / finalize_from_spec tests — which legitimately
reference internals (reshard_dataset, get_stats_hdf5, stream_stats_wds) —
stay in the sibling tests/pipeline/entrypoints/test_finalize_dataset_unit.py.
Extend both tests/_meta guards to cover the new canonical module; the
branch/unit module is deliberately not added to the entrypoint-only rail.

Test-only; every existing finalize test is preserved (26 -> 10 + 16).

Closes #1513
